feat: added test for authentication
