+++ api/assistant-api/api/assistant/get_assistant_authentication.go
@@ -30,13 +30,13 @@ func (assistantApi *assistantGrpcApi) GetAssistantAuthentication(
 		req.GetAssistantId(),
 	)
 	if err != nil {
-		assistantApi.logger.Errorf("failed to get assistant authentication: %v", err)
 		return utils.Error[protos.GetAssistantAuthenticationResponse](
 			err,
 			"Unable to get assistant authentication for the given assistant id.",
 		)
 	}
-	var out *protos.AssistantAuthentication
+	assistantApi.logger.Debugf("got assistant authentication for assistant id %+v", authConfig)
+	out := &protos.AssistantAuthentication{}
 	if err = utils.Cast(authConfig, out); err != nil {
 		assistantApi.logger.Errorf("unable to cast assistant authentication %v", err)
 	}
