+++ api/assistant-api/internal/services/assistant/authentication.impl.service.go
@@ -7,7 +7,6 @@ package internal_assistant_service
 
 import (
 	"context"
-	"errors"
 	"time"
 
 	internal_assistant_entity "github.com/rapidaai/api/assistant-api/internal/entity/assistants"
@@ -44,7 +43,9 @@ func (s *assistantAuthenticationService) Get(
 ) (*internal_assistant_entity.AssistantAuthentication, error) {
 	start := time.Now()
 	db := s.postgres.DB(ctx)
-	var out internal_assistant_entity.AssistantAuthentication
+
+	//
+	var out *internal_assistant_entity.AssistantAuthentication
 	tx := db.Preload("AssistantAuthenticationOption", "status = ?", type_enums.RECORD_ACTIVE).
 		Where("assistant_id = ? AND organization_id = ? AND project_id = ? AND status IN ?",
 			assistantId,
@@ -54,17 +55,13 @@ func (s *assistantAuthenticationService) Get(
 				type_enums.RECORD_ACTIVE,
 				type_enums.RECORD_INACTIVE,
 			}).
-		Order(clause.OrderByColumn{
-			Column: clause.Column{Name: "created_date"},
-			Desc:   true,
-		}).
-		First(&out)
-	if tx.Error != nil && !errors.Is(tx.Error, gorm.ErrRecordNotFound) {
+		Last(&out)
+	if tx.Error != nil {
 		s.logger.Benchmark("AssistantAuthenticationService.Get", time.Since(start))
 		return nil, tx.Error
 	}
 	s.logger.Benchmark("AssistantAuthenticationService.Get", time.Since(start))
-	return &out, nil
+	return out, nil
 }
 
 func (s *assistantAuthenticationService) Create(
